+++ engine/src/Core/Utils/Ranges.ts
@@ -29,6 +29,11 @@ export abstract class BaseRange {
     this.type = type;
   }
 
+  protected _resetPosition(x: number, y: number): void {
+    this.position.x = x;
+    this.position.y = y;
+  }
+
   /**
    * Check if the point is inside the range
    * @param point - the point to check in the range
@@ -47,18 +52,18 @@ export abstract class BaseRange {
 /**
  */
 export class Circle extends BaseRange {
+  radius;
+
   /**
    * Circle constructor, initialized position and radius
    * @param x - X coordinate of the position
    * @param y - Y coordinate of the position
    * @param radius - Circle's radius
    */
-  constructor(
-    x: number,
-    y: number,
-    readonly radius: number,
-  ) {
+  constructor(x: number, y: number, radius: number) {
     super(x, y, RangeType.circle);
+
+    this.radius = radius;
   }
 
   /**
@@ -78,29 +83,33 @@ export class Circle extends BaseRange {
   intersects(range: BaseRange): boolean {
     const pos1 = this.position,
       pos2 = range.position,
-      distPos = { x: Math.abs(pos2.x - pos1.x), y: Math.abs(pos2.y - pos1.y) },
-      r = this.radius;
+      r = this.radius,
+      dx = Math.abs(pos2.x - pos1.x),
+      dy = Math.abs(pos2.y - pos1.y);
 
     if (range instanceof Circle || range.type === (RangeType.circle as string)) {
       const circleRange = range as Circle,
         rSum = r + circleRange.radius,
-        dist = Math.hypot(distPos.x, distPos.y);
+        dist = Math.hypot(dx, dy);
 
       return rSum > dist;
     } else if (range instanceof Rectangle || range.type === (RangeType.rectangle as string)) {
       const rectRange = range as Rectangle,
         { width, height } = rectRange.size,
-        edges = Math.pow(distPos.x - width, squareExp) + Math.pow(distPos.y - height, squareExp);
+        edges = Math.pow(dx - width, squareExp) + Math.pow(dy - height, squareExp);
 
-      return (
-        edges <= r ** squareExp ||
-        (distPos.x <= r + width && distPos.y <= r + height) ||
-        distPos.x <= width ||
-        distPos.y <= height
-      );
+      return edges <= r ** squareExp || (dx <= r + width && dy <= r + height) || dx <= width || dy <= height;
     }
 
     return false;
+  }
+
+  reset(x: number, y: number, radius: number): this {
+    this._resetPosition(x, y);
+
+    this.radius = radius;
+
+    return this;
   }
 }
 
@@ -151,14 +160,25 @@ export class Rectangle extends BaseRange {
       return range.intersects(this);
     }
 
+    if (!(range instanceof Rectangle)) {
+      return false;
+    }
+
     const w = this.size.width,
       h = this.size.height,
       pos1 = this.position,
       pos2 = range.position,
-      size2 = range instanceof Rectangle ? range.size : { width: 0, height: 0 },
+      size2 = range.size,
       w2 = size2.width,
       h2 = size2.height;
 
     return pos2.x < pos1.x + w && pos2.x + w2 > pos1.x && pos2.y < pos1.y + h && pos2.y + h2 > pos1.y;
+  }
+
+  reset(x: number, y: number, width: number, height: number): this {
+    this._resetPosition(x, y);
+    this.size.width = width;
+    this.size.height = height;
+    return this;
   }
 }